Module pallet_mixnet::pallet::storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

CurrentSessionIndex

Index of the current session. This may be offset relative to the session index tracked by eg pallet_session; mixnet session indices are independent.

CurrentSessionStartBlock

Block in which the current session started.

Mixnodes

Mixnode sets by session index. Only the mixnode sets for the previous, current, and next sessions are kept; older sets are discarded.

NextAuthorityIds

Authority list for the next session.