Crate pallet_mixnet

This pallet is responsible for determining the current mixnet session and phase, and the mixnode set for each session.

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

Structs

BoundedMixnode

Like Mixnode, but encoded size is bounded.

Registration

A mixnode registration. A registration transaction is formed from one of these plus an AuthoritySignature.

Type Aliases

AuthorityIndex

Index of an authority in the authority list for a session.

BoundedMixnodeFor

BoundedMixnode type for the given configuration.

RegistrationFor

Registration type for the given configuration.