Type Alias pallet_mixnet::RegistrationFor

pub type RegistrationFor<T> = Registration<BlockNumberFor<T>, BoundedMixnodeFor<T>>;

Registration type for the given configuration.

Aliased Type

struct RegistrationFor<T> {
    pub block_number: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
    pub session_index: u32,
    pub authority_index: u32,
    pub mixnode: BoundedMixnode<BoundedVec<BoundedVec<u8, <T as Config>::MaxExternalAddressSize>, <T as Config>::MaxExternalAddressesPerMixnode>>,
}

Fields
block_number: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

Block number at the time of creation. When a registration transaction fails to make it on to the chain for whatever reason, we send out another one. We want this one to have a different hash in case the earlier transaction got banned somewhere; including the block number is a simple way of achieving this.

session_index: u32

The session during which this registration should be processed. Note that on success the mixnode is registered for the following session.

authority_index: u32

The index in the next session’s authority list of the authority registering the mixnode.

mixnode: BoundedMixnode<BoundedVec<BoundedVec<u8, <T as Config>::MaxExternalAddressSize>, <T as Config>::MaxExternalAddressesPerMixnode>>

Mixnode information to register for the following session.